Add abstract interface for disc image access.
The disc image being accessed used a FILE * directly. This is fine if
you only want to work with files on disc but if you need to use a
different interface to access the disc image data, this is not much use.

This change introduces a new discimage_t, through which the disc files
are accessed. This means that the access to the disc image can be changed
from being a file to being (say) a block device accessed through a
different interface.

All the internal operations that read or wrote through the standard
FILE * stream interfaces, for the disc image, have now been replaced
with operations that use the discimage_t methods.